src_device/pyproject.toml
[project]
name         = "edge-device"
description  = "Internet of Things Device"
authors      = [{name = "Joshoua Bigler"}]
version      = "0.0.1"
dependencies = [
  "python-dotenv>=1.0.1",
  "grpcio==1.68.1",
  "grpcio-tools==1.68.1",
  "iot-libs>=0.0.8",
  "pandas>=2.2.3",
  "statsmodels>=0.14.4",
  "websockets>=13.1",
  "requests>=2.32.3"
]
[tool.setuptools]
package-dir  = { "" = "." }
packages = { find = { include = ["edge_device*"] } } 
[tool.pytest.ini_options]
testpaths = ["test"]